Update EYESY modes and add new features
- Add new custom modes (dancing_animals, dancing_character, musical_staff)
- Add new example modes (Joy Division, ZACH Reactive, Surf Waves, Rain, Snow, etc.)
- Update existing scope and trigger modes
- Add tools directory with eyesy_runner.py test runner
- Update documentation (API_REFERENCE, QUICK_REFERENCE)
- Add .cursorrules for project conventions
- Update README with testing instructions

+++ examples/scopes/S - Dancing Circle - Image/main.py
@@ -4,20 +4,17 @@ import time
 import random
 import glob
 import math
-
 #Knob1 - image size
 #Knob2 - diameter of 'dance'
 #Knob3 - circle size & unfilled/filled option
 #Knob4 - foreground color
 #Knob5 - background color
-
 last_point = [0, 360]
 i = 0
 images = []
 image_index = 0
 lx = 0
 ly = 0
-
 def setup(screen, eyesy):
     global images, xr, yr, last_point
     xr = eyesy.xres
@@ -28,7 +25,6 @@ def setup(screen, eyesy):
         print('loading image file: ' + filename)
         img = pygame.image.load(filepath).convert_alpha()
         images.append(img)
-
 def draw(screen, eyesy):
     global last_point, images, i, lx, ly, xr,yr
     eyesy.color_picker_bg(eyesy.knob5)
@@ -40,12 +36,10 @@ def draw(screen, eyesy):
     y1 = int(eyesy.knob2 * yr) + ((eyesy.audio_in[i]*0.00003058) *(yr/2))
     x = i * (xr/100)
     color = eyesy.color_picker_lfo(eyesy.knob4)
-
     R = (eyesy.knob2*x300)-(x300)
     R = R + ((eyesy.audio_in[i]*0.00003058)*(yr/2))
     x = R * math.cos((i /  100.) * 6.28) + (xr/2)
     y = R * math.sin((i /  100.) * 6.28) + (yr/2)
-    
     max_circle = x300
     image_size = 1
     circle_size = 0
@@ -54,14 +48,12 @@ def draw(screen, eyesy):
         circle_size = int(eyesy.knob3 * max_circle)
         line_width = 0
     if eyesy.knob3 >.501 :
-        circle_size = abs(max_circle-int(eyesy.knob3 * max_circle)) 
+        circle_size = abs(max_circle-int(eyesy.knob3 * max_circle))
         line_width =  abs(x30-int(eyesy.knob3 * x30))
-    
     pygame.draw.circle(screen,color,(int(x),int(y)), circle_size, line_width)
     image = images[0]
     image_w = int(image.get_width() * eyesy.knob1*6)
     image_h = int(image.get_height() * eyesy.knob1*6)
     image = pygame.transform.scale(image, (image_w, image_h))
-    screen.blit(image, (x-(image_w/2), y-(image_h/2)))
-    
+    screen.blit(image, (int(x-(image_w/2)), int(y-(image_h/2))))
     i = (i + 1) % 100
